Distributed DDL queries

+++ dbms/src/Interpreters/InterpreterDropQuery.cpp
@@ -6,6 +6,7 @@
 #include <Interpreters/InterpreterDropQuery.h>
 #include <Storages/IStorage.h>
 #include <Databases/IDatabase.h>
+#include <Interpreters/DDLWorker.h>
 
 
 namespace DB
@@ -27,10 +28,13 @@ InterpreterDropQuery::InterpreterDropQuery(const ASTPtr & query_ptr_, Context & 
 
 BlockIO InterpreterDropQuery::execute()
 {
+    ASTDropQuery & drop = typeid_cast<ASTDropQuery &>(*query_ptr);
+
+    if (!drop.cluster.empty())
+        return executeDDLQueryOnCluster(query_ptr, context);
+
     String path = context.getPath();
     String current_database = context.getCurrentDatabase();
-
-    ASTDropQuery & drop = typeid_cast<ASTDropQuery &>(*query_ptr);
 
     bool drop_database = drop.table.empty() && !drop.database.empty();
